posts populated by database

+++ Makefile
@@ -21,6 +21,6 @@ stack:
 
 clean:
 	@echo "Cleaning up"
-	docker stop $(shell docker ps -aq)
-	docker rm $(shell docker ps -aq)
-	docker rmi $(shell docker images -aq)
+	if [ "`docker ps -aq`" ]; then docker stop $(shell docker ps -aq); fi
+	if [ "`docker ps -aq`" ]; then docker rm $(shell docker ps -aq); fi
+	if [ "`docker images -aq`" ]; then docker rmi $(shell docker images -aq); fi
